listing-edit-and-delete-polls-after-updating-poll-pages
